split components, fix display logic
