+++ components/shared/atoms/Button/Button.tsx
@@ -1,0 +1,3 @@
+import { Button } from './styled';
+
+export default Button;
